requirements.txt
Flask==2.2.5
numpy==1.24.3
Pillow==10.1.0
opencv-python==4.8.0.76
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
scipy==1.10.1
scikit-image==0.19.3
scikit-learn==1.2.2
requests==2.31.0
fastapi==0.103.1
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.12
